Move psi into psi:pack folder
