+++ components/psi/pack/src/test/resources/DataObjectIconPack.kt
