+++ components/psi/pack/src/test/resources/NestedIconPack.kt
