+++ components/psi/pack/src/test/resources/SimpleIconPack.kt
