+++ gradle/libs.versions.toml
@@ -7,7 +7,7 @@ kotlin = "2.0.10"
 androidx-activity-compose = "androidx.activity:activity-compose:1.9.1"
 android-build-tools = "com.android.tools:sdk-common:31.5.2"
 
-highlights = "dev.snipme:highlights:0.9.0"
+highlights = "dev.snipme:highlights:0.9.1"
 
 koin-compose = "io.insert-koin:koin-compose:4.0.0-RC1"
