+++ idea-plugin/build.gradle.kts
@@ -10,7 +10,7 @@ dependencies {
     implementation(projects.components.generator.iconpack)
     implementation(projects.components.generator.imagevector)
     implementation(projects.components.parser)
-    implementation(projects.components.psi)
+    implementation(projects.components.psi.pack)
 
     compileOnly(compose.desktop.currentOs)
     implementation(compose.desktop.common)
