+++ settings.gradle.kts
@@ -52,5 +52,5 @@ include("components:generator:common")
 include("components:generator:iconpack")
 include("components:generator:imagevector")
 include("components:parser")
-include("components:psi")
+include("components:psi:pack")
 include("playground:app")
